+++ app/src/androidTest/java/org/simple/clinic/benchmark/overdue/OverdueBenchmark.kt
@@ -34,7 +34,7 @@ class OverdueBenchmark : BaseBenchmarkTest() {
   @Test
   fun searching_overdue_patients_in_a_facility() {
     val pagingTestCase = PagingTestCase(
-        pagingSource = appointmentRepository.searchOverduePatient(
+        pagingSource = appointmentRepository.searchOverduePatient_Old(
             searchQuery = "Ani",
             since = LocalDate.now(),
             facilityId = UUID.fromString("c68603b3-9293-4783-bd76-0dc425c0c5d2")