+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Bump fragment to v1.5.1
 - Show chip input text view in overdue search when overdue v2 search is enabled
+- Add overdue search query for supporting multiple search inputs
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-08-01-8350